+++ Evolution_1/Living_Beings/plant.cpp
@@ -312,7 +312,7 @@ Plant* Plant::reproduction(){
     std::map<Enum_parameters, double> param_new_plant;
     for ( Enum_parameters param = (Enum_parameters)0; param != last; param=(Enum_parameters)(param+1) ) {
         double val = normal_distrib(parameters[param],0.1);
-        param_new_plant.insert(std::pair<Enum_parameters, float>(param, val));
+        param_new_plant.insert(std::pair<Enum_parameters, double>(param, val));
     }
     Plant* p = new Plant(param_new_plant, this->get_scene());
     std::random_device rd;
